requirements.txt
# Add your requirements here like:
click
xmltodict
pandas
python-daemon
schedule